refactor : Validation class 생성 및 exception 처리 분리
예외처리 클래스 생성 후 분리.
